Refactor: refactoring component directory
