+++ src/App.tsx
@@ -5,7 +5,7 @@ import { TimerContextProvider } from "./components/Provider/TimerContextProvider
 
 import Home from "./pages/Home/Home";
 import QuizPage from "./pages/QuizPage/QuizPage";
-import QuizResult from "./pages/QuizResult/QuizResult";
+import QuizResultPage from "./pages/QuizResult/QuizResultPage";
 import Layout from "./components/Layout/Layout";
 
 function App() {
@@ -16,7 +16,7 @@ function App() {
           <Routes>
             <Route path="/" element={<Home data-cy="homeComponent" />} />
             <Route path="/quiz" element={<QuizPage />} />
-            <Route path="/result" element={<QuizResult />} />
+            <Route path="/result" element={<QuizResultPage />} />
           </Routes>
         </Layout>
       </RecoilRoot>
