+++ src/components/Quiz/QuizResultComponent.module.css
@@ -41,7 +41,7 @@
 
 .quiz_result_table_box {
   display: inline-block;
-  margin-top: 50px;
+  margin-top: 30px;
 }
 
 .quiz_result_table {
