+++ src/components/Quiz/QuizStartComponent.module.css
@@ -1,0 +1,31 @@
+.home_wrapper {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  height: 100%;
+  text-align: center;
+}
+
+.home_title_box {
+  margin-top: 20px;
+}
+
+.home_logo_image_box {
+  overflow: hidden;
+}
+
+.home_logo_image {
+  width: 300px;
+  height: 300px;
+}
+
+.bold_tag {
+  color: rgb(100, 149, 237);
+}
+
+.footer_tag {
+  margin-top: 20px;
+  font-size: 14px;
+  font-weight: 500;
+  color: gray;
+}
