--- src/pages/QuizPage/QuizPage.module.css
@@ -1,33 +1,0 @@
-.quiz_page_wrapper {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  height: 100%;
-}
-
-.footer_box {
-  display: flex;
-
-  flex-direction: column;
-}
-
-.footer_contents {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-top: 20px;
-}
-
-.timer {
-  font-size: 16px;
-  font-weight: 500;
-  color: gray;
-}
-
-.question_length {
-  display: inline-block;
-  padding: 10px 16px;
-  border-radius: 16px;
-  border: 4px solid rgb(197, 217, 255);
-  font-weight: 600;
-}
